+++ src/main/java/com/groom/e_commerce/product/domain/entity/Product.java
@@ -67,7 +67,7 @@ public class Product extends BaseEntity {
 	private Boolean hasOptions;
 
 	@Column(name = "price", precision = 12, scale = 2)
-	private BigDecimal price;
+	private Long price;
 
 	@Column(name = "stock_quantity")
 	private Integer stockQuantity;
@@ -87,7 +87,7 @@ public class Product extends BaseEntity {
 
 	@Builder
 	public Product(UUID ownerId, Category category, String title, String description,
-		String thumbnailUrl, Boolean hasOptions, BigDecimal price, Integer stockQuantity) {
+		String thumbnailUrl, Boolean hasOptions, Long price, Integer stockQuantity) {
 		this.ownerId = ownerId;
 		this.category = category;
 		this.title = title;
@@ -100,7 +100,7 @@ public class Product extends BaseEntity {
 	}
 
 	public void update(Category category, String title, String description, String thumbnailUrl,
-		BigDecimal price, Integer stockQuantity, ProductStatus status) {
+		Long price, Integer stockQuantity, ProductStatus status) {
 		if (category != null) {
 			this.category = category;
 		}
